db_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-jose[cryptography]